Implement Naming of Serialized Fields (#61)
* Implement Naming of Serialized Fields

* Fix typo

* WIP

* Not sure

* I think this fixes the build errors

* Maybe this time it will work
