+++ parameter_tools/src/main/java/coppercore/parameter_tools/JSONName.java
@@ -1,0 +1,12 @@
+package coppercore.parameter_tools;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.FIELD)
+public @interface JSONName {
+    String value() default "No name given";
+}
