+++ parameter_tools/src/main/java/coppercore/parameter_tools/JSONSync.java
@@ -1,11 +1,9 @@
 package coppercore.parameter_tools;
 
 import com.google.gson.ExclusionStrategy;
-import com.google.gson.FieldAttributes;
-import com.google.gson.FieldNamingPolicy;
+import com.google.gson.FieldNamingStrategy;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import com.google.gson.LongSerializationPolicy;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.FileWriter;
@@ -17,6 +15,7 @@ import java.io.IOException;
  */
 public class JSONSync<T> {
 
+    private final JSONSyncConfig defaultConfig = new JSONSyncConfigBuilder().build();
     private final Gson gson; // Gson instance for serialization and deserialization
     private T instance; // The object being synchronized
     private String file; // File path for the JSON file
@@ -86,28 +85,17 @@ public class JSONSync<T> {
      * @return A configured Gson instance.
      */
     private Gson generateGson() {
-        ExclusionStrategy strategy =
-                new ExclusionStrategy() {
-                    @Override
-                    public boolean shouldSkipField(FieldAttributes field) {
-                        return (field.getAnnotation(JSONExclude.class) != null);
-                    }
-
-                    @Override
-                    public boolean shouldSkipClass(Class<?> clazz) {
-                        return false;
-                    }
-                };
-
+        ExclusionStrategy jsonExcludeStrategy = new JSONExcludeExclusionStrategy();
+        FieldNamingStrategy jsonNameStrategy = new JSONNamingStrategy(this.config.namingPolicy());
         GsonBuilder builder = new GsonBuilder();
-        if (this.config.serializeNulls) builder.serializeNulls();
-        if (this.config.prettyPrinting) builder.setPrettyPrinting();
-        if (this.config.excludeFieldsWithoutExposeAnnotation)
+        if (this.config.serializeNulls()) builder.serializeNulls();
+        if (this.config.prettyPrinting()) builder.setPrettyPrinting();
+        if (this.config.excludeFieldsWithoutExposeAnnotation())
             builder.excludeFieldsWithoutExposeAnnotation();
-        builder.setFieldNamingPolicy(this.config.namingPolicy)
-                .setLongSerializationPolicy(this.config.longSerializationPolicy)
-                .addDeserializationExclusionStrategy(strategy)
-                .addSerializationExclusionStrategy(strategy);
+        builder.setFieldNamingStrategy(jsonNameStrategy)
+                .setLongSerializationPolicy(this.config.longSerializationPolicy())
+                .addDeserializationExclusionStrategy(jsonExcludeStrategy)
+                .addSerializationExclusionStrategy(jsonExcludeStrategy);
         return builder.create();
     }
 
@@ -138,98 +126,6 @@ public class JSONSync<T> {
             return new FileWriter(path);
         } catch (IOException e) {
             throw new RuntimeException("IOException " + path, e);
-        }
-    }
-
-    /** Configuration class for customizing Gson behavior. */
-    private static record JSONSyncConfig(
-            boolean serializeNulls,
-            boolean prettyPrinting,
-            boolean excludeFieldsWithoutExposeAnnotation,
-            FieldNamingPolicy namingPolicy,
-            LongSerializationPolicy longSerializationPolicy) {
-        public JSONSyncConfig(JSONSyncConfigBuilder builder) {
-            this(
-                    builder.serializeNulls,
-                    builder.prettyPrinting,
-                    builder.excludeFieldsWithoutExposeAnnotation,
-                    builder.namingPolicy,
-                    builder.longSerializationPolicy);
-        }
-    }
-
-    /** Builder class for creating a JSONSyncConfig instance. */
-    public static class JSONSyncConfigBuilder {
-        public boolean serializeNulls = false;
-        public boolean prettyPrinting = false;
-        public boolean excludeFieldsWithoutExposeAnnotation = false;
-        public FieldNamingPolicy namingPolicy = FieldNamingPolicy.IDENTITY;
-        public LongSerializationPolicy longSerializationPolicy = LongSerializationPolicy.DEFAULT;
-
-        /**
-         * Sets whether null fields should be serialized.
-         *
-         * @param serializeNulls True to serialize null fields, false otherwise.
-         * @return The builder instance.
-         */
-        public JSONSyncConfigBuilder setSerializeNulls(boolean serializeNulls) {
-            this.serializeNulls = serializeNulls;
-            return this;
-        }
-
-        /**
-         * Sets whether the JSON output should use pretty printing.
-         *
-         * @param prettyPrinting True to enable pretty printing, false otherwise.
-         * @return The builder instance.
-         */
-        public JSONSyncConfigBuilder setPrettyPrinting(boolean prettyPrinting) {
-            this.prettyPrinting = prettyPrinting;
-            return this;
-        }
-
-        /**
-         * Sets whether fields without @Expose annotations should be excluded.
-         *
-         * @param excludeFieldsWithoutExposeAnnotation True to exclude fields, false otherwise.
-         * @return The builder instance.
-         */
-        public JSONSyncConfigBuilder setExcludeFieldsWithoutExposeAnnotation(
-                boolean excludeFieldsWithoutExposeAnnotation) {
-            this.excludeFieldsWithoutExposeAnnotation = excludeFieldsWithoutExposeAnnotation;
-            return this;
-        }
-
-        /**
-         * Sets the naming policy for fields in the JSON output.
-         *
-         * @param namingPolicy The field naming policy.
-         * @return The builder instance.
-         */
-        public JSONSyncConfigBuilder setNamingPolicy(FieldNamingPolicy namingPolicy) {
-            this.namingPolicy = namingPolicy;
-            return this;
-        }
-
-        /**
-         * Sets the serialization policy for long values.
-         *
-         * @param longSerializationPolicy The serialization policy for long values.
-         * @return The builder instance.
-         */
-        public JSONSyncConfigBuilder setLongSerializationPolicy(
-                LongSerializationPolicy longSerializationPolicy) {
-            this.longSerializationPolicy = longSerializationPolicy;
-            return this;
-        }
-
-        /**
-         * Builds the configuration object.
-         *
-         * @return A JSONSyncConfig instance.
-         */
-        public JSONSyncConfig build() {
-            return new JSONSyncConfig(this);
         }
     }
 }
